requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==1.26.4
opencv-python==4.9.0.80
pillow==11.3.0
scikit-image==0.24.0
scipy==1.13.1
tifffile==2024.8.30
torch==2.2.2
torchvision==0.17.2
Werkzeug==3.1.3